+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,13 @@
 # versa-storybook
 
+## 0.0.28
+
+### Patch Changes
+
+- Updated dependencies
+  - @versaprotocol/examples@1.8.1
+  - @versaprotocol/react@1.8.1
+
 ## 0.0.27
 
 ### Patch Changes